backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
langchain_openai
langchain_community
faiss-cpu
json
